requirements.txt
justpy
python-dateutil
jinja2
addict
pdfminer.six